more robust against a network failure
git-svn-id: https://groke.otago.ac.nz/ant/ant@1109 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/atire_api_remote.h
@@ -43,7 +43,7 @@ public:
 	/*
 		Load a new doclist and index by filename.
 	*/
-	virtual int load_index(char *doclist_filename, char *index_filename);
+	virtual long load_index(char *doclist_filename, char *index_filename);
 
 	/*
 		Return an XML document describing the index that is currently loaded (incl filenames).
